Commit before i change invoice

+++ BackEnd/QArte/QArte.Services/DTOs/FeeDTO.cs
@@ -5,9 +5,8 @@ namespace QArte.Services.DTOs
 	public class FeeDTO
 	{
 		public int ID { get; set; }
-		public decimal Amount { get; set; }
+		public long Amount { get; set; }
 		public string Currency { get; set; }
-		public decimal ExchangeRate { get; set; }
+		public double ExchangeRate { get; set; }
     }
 }
-
